adicionado autorização no modelo crud

+++ src/actions/users.ts
@@ -1,3 +1,5 @@
+import { ErrorsMessages } from "@/config/messages";
+import { ROLES } from "@/enums/roles";
 import { prisma } from "@/lib/prisma";
 import { DefaultReturn, ErrorReturn } from "@/types/actions/_general";
 import {
@@ -6,6 +8,7 @@ import {
   UpdateUser as UpdateResource,
 } from "@/types/actions/users";
 import { User as Model } from "@prisma/client";
+import { getServerSession } from "next-auth";
 
 const model = prisma.user;
 //Para usar como modelo basta trocar CreateUser, UpdateUser e User
@@ -14,6 +17,12 @@ export async function create(
   data: CreateResource
 ): Promise<DefaultReturn<Model> | ErrorReturn> {
   try {
+    const session = await getServerSession();
+
+    if (session?.user.role !== ROLES.ADMIN) {
+      return { error: ErrorsMessages.not_authorized };
+    }
+
     const response = await model.create({ data });
     return { data: response };
   } catch (error) {
@@ -24,6 +33,11 @@ export async function create(
 
 export async function find(): Promise<ListReturn | ErrorReturn> {
   try {
+    const session = await getServerSession();
+
+    if (session?.user.role !== ROLES.ADMIN) {
+      return { error: ErrorsMessages.not_authorized };
+    }
     const response = await model.findMany();
     return { data: response };
   } catch (error) {
@@ -36,6 +50,11 @@ export async function update(
   data: UpdateResource
 ): Promise<DefaultReturn<Model> | ErrorReturn> {
   try {
+    const session = await getServerSession();
+
+    if (session?.user.role !== ROLES.ADMIN) {
+      return { error: ErrorsMessages.not_authorized };
+    }
     const { id, ...dataToUpdate } = data;
     const response = await model.update({
       where: { id },
@@ -50,6 +69,11 @@ export async function update(
 
 export async function remove(id: Model["id"]): Promise<void | ErrorReturn> {
   try {
+    const session = await getServerSession();
+
+    if (session?.user.role !== ROLES.ADMIN) {
+      return { error: ErrorsMessages.not_authorized };
+    }
     await model.delete({
       where: { id },
     });
